Tidy up test layout
- Unify all tests/ into a single file to reduce the number of output
  targets during `cargo test`
- This also avoids double-compiling the normalization_tests.rs
- Move normalization_tests.rs to be a child of tests/data/ to avoid
  getting compiled as a test unit, and further removing useless output
  from cargo test
